backend/pyproject.toml
[tool.poetry]
name = "agui-mcp-voice-demo"
version = "0.1.0"
description = "Open-source demo: AG-UI + LangGraph ReAct + MCP + Voice"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.14"

# LLM + agent
langchain-openai = ">=0.3.0,<1.0.0"
langgraph = ">=0.4.0,<2.0.0"
langgraph-checkpoint-sqlite = ">=3.0.0,<4.0.0"

# AG-UI adapter (LangGraph → SSE)
ag-ui-langgraph = ">=0.0.25,<1.0.0"

# MCP
langchain-mcp-adapters = ">=0.1.0,<1.0.0"
mcp = ">=1.0.0,<2.0.0"

# FastAPI
fastapi = ">=0.115.0,<1.0.0"
uvicorn = {extras = ["standard"], version = ">=0.32.0,<1.0.0"}
python-multipart = ">=0.0.12,<1.0.0"

# HTTP clients
httpx = ">=0.28.0,<1.0.0"
aiohttp = ">=3.11.0,<4.0.0"

# Utilities
python-dotenv = ">=1.0.0,<2.0.0"
aiosqlite = ">=0.20.0,<1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0.0"
pytest-asyncio = ">=0.24.0"
httpx = ">=0.28.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
